fix: refactored validaion and options, set button name props as unnecessary, refactor AdminTails.tsx structure
