+++ src/layout/Button/Button.tsx
@@ -3,7 +3,7 @@ import * as React from 'react';
 import styles from './Button.module.scss';
 
 interface IButtonProps {
-	name: string;
+	name?: string;
 	btnClasses: string;
 	onClick?: () => void;
 	type?: 'button' | 'submit' | 'reset';
