+++ src/pages/AdminPhotos/AdminPhotos.module.scss
@@ -6,7 +6,7 @@
 	flex-direction: column;
 	gap: 32px;
 	padding: 15px;
-	width: 100%;
+
 }
 
 .title {
@@ -21,6 +21,7 @@
   margin-bottom: 10px;
   height: 80px;
   overflow: hidden;
+  width: 30%;
 }
 
 .inputContainer {
