+++ src/pages/AdminTails/TailsList.tsx
@@ -1,10 +1,9 @@
 import * as React from 'react';
-import { FaTrash, FaEdit, FaRegPlusSquare } from 'react-icons/fa';
+import { FaTrash, FaEdit } from 'react-icons/fa';
 import { useEffect, useState } from 'react';
-import ArrowIconUp from '../../assets/dropdown_arrow_up.svg';
-import ArrowIconDown from '../../assets/dropdown_arrow_down.svg';
+
 import { UseQueryResult } from '@tanstack/react-query';
-import { AdminTailsData, TailsListData } from './AdminTails';
+import { AdminTailsData } from './AdminTails';
 import styles from '../../pages/AdminTails/TailsList.module.scss';
 import Tail from '../../components/Tail/Tail';
 import Button from '../../layout/Button/Button';
@@ -12,14 +11,6 @@ import Button from '../../layout/Button/Button';
 interface TailsListProps {
 	data: UseQueryResult<AdminTailsData, Error>;
 	changeShowForm: (a: boolean, b: string, c: number) => void;
-}
-
-interface CustomStyles {
-	control?: (provided: any, state: any) => any;
-	dropdownIndicator?: (provided: any) => any;
-	indicatorSeparator?: () => any;
-	menu?: (provided: any) => any;
-	option?: (provided: any) => any;
 }
 
 
@@ -53,58 +44,6 @@ const handleEdit = (id) => {
 	const handleDelete = () => {
 		console.log('Delete');
 
-	};
-
-	const customStyles: CustomStyles = {
-		control: (provided, state) => ({
-			...provided,
-			padding: '6px 16px',
-			borderRadius: '40px',
-			borderColor: '#b6e1f2',
-			color: '#858585',
-			fontSize: '20px',
-			fontWeight: 500,
-			cursor: 'pointer',
-			position: 'relative',
-			'&:hover': {
-				borderColor: '#b6e1f2',
-			},
-			'&:before': {
-				content: '""',
-				backgroundImage: state.menuIsOpen ? `url(${ArrowIconUp})` : `url(${ArrowIconDown})`,
-				backgroundRepeat: 'no-repeat',
-				backgroundPosition: 'center',
-				backgroundSize: '24px 24px',
-				width: '24px',
-				height: '24px',
-				position: 'absolute',
-				right: '8px',
-				top: '50%',
-				transform: 'translateY(-50%)',
-			},
-		}),
-		indicatorSeparator: () => ({
-			display: 'none',
-		}),
-		dropdownIndicator: (provided) => ({
-			...provided,
-			display: 'none',
-		}),
-		menu: (provided) => ({
-			...provided,
-			borderRadius: '24px',
-			boxShadow: '0 0 0 1px #b6e1f2',
-		}),
-		option: (provided) => ({
-			...provided,
-			backgroundColor: 'white',
-			cursor: 'pointer',
-			color: '#0D0031',
-			borderRadius: '24px',
-			'&:hover': {
-				color: '#009cd9',
-			},
-		}),
 	};
 
 
@@ -145,7 +84,7 @@ const handleEdit = (id) => {
 										<div className={styles.btnContainer}>
 											<div className={styles.btnBox}>
 												<Button
-													onClick={() => console.log('delete')}
+													onClick={handleDelete}
 													type={'button'}
 													btnClasses={'delete'}  children={<FaTrash  className={styles.deleteIcon}/>}/>
 											</div>
